Renames nextTrips to driverTrips
