+++ app/src/driver/java/mainFunctionality/driverTrips/TripsAdapter.java
@@ -1,4 +1,4 @@
-package mainFunctionality.nextTrips;
+package mainFunctionality.driverTrips;
 
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
